+++ src/filter/range.rs
@@ -1,8 +1,11 @@
 use super::*;
 
-/// NOTE: le and lt operations are written for big-endian format of keys
 #[derive(Debug, Default, Serialize, Deserialize, Clone)]
-pub struct RangeFilter<K: Key> {
+/// Range filter
+pub struct RangeFilter<K>
+where
+    for<'a> K: Key<'a>,
+{
     #[serde(serialize_with = "serialize_key", deserialize_with = "deserialize_key")]
     min: K,
     #[serde(serialize_with = "serialize_key", deserialize_with = "deserialize_key")]
@@ -13,7 +16,7 @@ pub struct RangeFilter<K: Key> {
 #[async_trait::async_trait]
 impl<K> FilterTrait<K> for RangeFilter<K>
 where
-    K: Key,
+    for<'a> K: Key<'a>,
 {
     fn add(&mut self, key: &K) {
         self.add(key);
@@ -32,9 +35,16 @@ where
         self.add(&other.max);
         true
     }
+
+    fn clear_filter(&mut self) {
+        self.clear()
+    }
 }
 
-impl<K: Key> RangeFilter<K> {
+impl<K> RangeFilter<K>
+where
+    for<'a> K: Key<'a>,
+{
     /// Create filter
     pub fn new() -> Self {
         Self {
@@ -77,15 +87,24 @@ impl<K: Key> RangeFilter<K> {
     }
 }
 
-fn serialize_key<K: Key, S: serde::Serializer>(key: &K, serializer: S) -> Result<S::Ok, S::Error> {
+fn serialize_key<K, S>(key: &K, serializer: S) -> Result<S::Ok, S::Error>
+where
+    for<'a> K: Key<'a>,
+    S: serde::Serializer,
+{
     serializer.serialize_bytes(key.as_ref())
 }
 
-fn deserialize_key<'de, K: Key, D: serde::Deserializer<'de>>(
-    deserializer: D,
-) -> Result<K, D::Error> {
+fn deserialize_key<'de, K, D>(deserializer: D) -> Result<K, D::Error>
+where
+    for<'a> K: Key<'a>,
+    D: serde::Deserializer<'de>,
+{
     struct KeyVisitor<K>(PhantomData<K>);
-    impl<'de, K: Key> serde::de::Visitor<'de> for KeyVisitor<K> {
+    impl<'de, K> serde::de::Visitor<'de> for KeyVisitor<K>
+    where
+        for<'a> K: Key<'a>,
+    {
         type Value = K;
 
         fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
@@ -105,7 +124,7 @@ fn deserialize_key<'de, K: Key, D: serde::Deserializer<'de>>(
 
 #[cfg(test)]
 mod tests {
-    use crate::Key;
+    use crate::{Key, RefKey};
 
     use super::RangeFilter;
 
@@ -113,6 +132,15 @@ mod tests {
 
     #[derive(Debug, Clone, Default, PartialEq, Eq, PartialOrd, Ord)]
     struct ProperKey(Vec<u8>);
+
+    #[derive(PartialEq, Eq, PartialOrd, Ord)]
+    struct RefProperKey<'a>(&'a [u8]);
+
+    impl<'a> From<&'a [u8]> for RefProperKey<'a> {
+        fn from(v: &'a [u8]) -> Self {
+            Self(v)
+        }
+    }
 
     impl AsRef<[u8]> for ProperKey {
         fn as_ref(&self) -> &[u8] {
@@ -127,12 +155,19 @@ mod tests {
         }
     }
 
-    impl Key for ProperKey {
+    impl<'a> RefKey<'a> for RefProperKey<'a> {}
+
+    impl<'a> Key<'a> for ProperKey {
         const LEN: u16 = LEN;
+
+        type Ref = RefProperKey<'a>;
     }
 
     #[derive(Debug, Clone, Default, PartialEq, Eq)]
     struct WrongKey(Vec<u8>);
+
+    #[derive(PartialEq, Eq, PartialOrd, Ord)]
+    struct RefWrongKey<'a>(&'a [u8]);
 
     impl AsRef<[u8]> for WrongKey {
         fn as_ref(&self) -> &[u8] {
@@ -159,8 +194,18 @@ mod tests {
         }
     }
 
-    impl Key for WrongKey {
+    impl<'a> From<&'a [u8]> for RefWrongKey<'a> {
+        fn from(v: &'a [u8]) -> Self {
+            Self(v)
+        }
+    }
+
+    impl<'a> RefKey<'a> for RefWrongKey<'a> {}
+
+    impl<'a> Key<'a> for WrongKey {
         const LEN: u16 = LEN;
+
+        type Ref = RefWrongKey<'a>;
     }
 
     fn to_key<K: From<Vec<u8>>>(i: usize) -> K {