+++ src/record.rs
@@ -90,7 +90,10 @@ impl Record {
     }
 
     /// Creates new `Record` with provided data, key and meta.
-    pub fn create<K: Key>(key: &K, data: Vec<u8>, meta: Meta) -> bincode::Result<Self> {
+    pub fn create<K>(key: &K, data: Vec<u8>, meta: Meta) -> bincode::Result<Self>
+    where
+        for<'a> K: Key<'a>,
+    {
         let key = key.as_ref().to_vec();
         let meta_size = meta.serialized_size()?;
         let data_checksum = CRC32C.checksum(&data);
@@ -114,7 +117,10 @@ impl Record {
         Ok(buf)
     }
 
-    pub(crate) fn deleted<K: Key>(key: &K) -> bincode::Result<Self> {
+    pub(crate) fn deleted<K>(key: &K) -> bincode::Result<Self>
+    where
+        for<'a> K: Key<'a> + 'static,
+    {
         let mut record = Record::create(key, vec![], Meta::default())?;
         record.header.mark_as_deleted()?;
         Ok(record)
@@ -234,6 +240,7 @@ impl Header {
         self.to_raw().map(|raw| CRC32C.checksum(&raw))
     }
 
+    /// Used for migration
     pub(crate) fn with_reversed_key_bytes(mut self) -> bincode::Result<Self> {
         self.key.reverse();
         self.update_checksum()?;