+++ CHANGELOG.md
@@ -17,6 +17,54 @@ Pearl changelog
 #### Updated
 
 
+
+## [0.13.0] - 2022-08-04
+#### Added
+- Add method to receive occupied disk space (#170)
+
+#### Changed
+- Defer index drop on delete (#165)
+
+#### Fixed
+- Restore documentation publishing (#173)
+
+#### Updated
+
+
+## [0.12.0] - 2022-05-23
+#### Added
+- Add hierarchical filters support (#126)
+- Add tools to recover and migrate blobs and indexes (#148)
+- Add blob size to index header (#153)
+- Add magic byte to index header (#152)
+
+#### Changed
+
+#### Fixed
+- Fix delete with no active blob set (#167)
+
+#### Updated
+
+
+## [0.11.0] - 2022-04-18
+#### Added
+Add support for hierarchical range filters (#154)
+
+#### Changed
+- Remove active checking for blob update (#159)
+
+#### Fixed
+- corrupted blob now should be saved in case of 'unexpected eof' error (#160)
+
+#### Updated
+
+
+## [0.10.0] - 2022-03-30
+#### Changed
+- Add refkey trait (#141)
+- Change nightly to stable rust toolchain
+
+
 ## [0.9.2] - 2022-02-14
 #### Added
 - Add record delete api method (#103)